+++ lib/core/loader.ts
@@ -1,12 +1,12 @@
 import path from "path";
-import { Constructor } from "../@types/types";
+import { Constructor } from "../types/types";
 import { promises as fs } from "fs";
 
 const promises: Promise<void>[] = [];
 
 export const loader = {
   async load(paths: string[]) {
-    injectViaAbsPath(paths);
+    await injectViaAbsPath(paths);
   },
 };