+++ lib/index.ts
@@ -11,3 +11,5 @@ export * from "./core/Context";
 export * from "./interface";
 
 export { Interceptor, ExceptionFilter } from "./core/middlewate-storage";
+
+export { NotFoundException, ParamsNotInvalideException } from "./constant/Exception";