+++ lib/interface/index.ts
@@ -1,3 +1,3 @@
 export { Next } from "koa";
 
-export * from "../@types/types";
+export * from "../types/types";