+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-[2020/09/09] v0.1.0
+[2020/09/09] v0.1.6 //undo
+
+- remove metadata-storage when app run
+
+[2020/09/09] v0.1.5
+
+- modify readme
+- change internal interceptor & exception filter
+  [2020/09/09] v0.1.4
+
+- fix dynamic import bug
+  [2020/09/09] v0.1.3
+
+- change useGlobalMiddleware signatrue
+  [2020/09/09] v0.1.2
+
+- change lib/@types -> lib/types
+  [2020/09/09] v0.1.0
 
 - life circle
 - middleware-storage